fix: cast logid to int before calling now natively-typed helpers
slowlog_has_stats_cache(int \) / slowlog_get_chart_object_live(int \, ...)
were called with \ = get_filter_request_var('logid') as-is. In the Pest test
bootstrap that stub returns '' (empty string), which isn't a valid numeric
string for weak-mode int coercion, so PHP threw a hard TypeError instead of
coercing - breaking 3 Unit tests in CI. Cast \ to (int) at both retrieval
points (slowlog_get_chart_object() and slowlog_get_stats_chart_object()).

+++ slowlog_functions.php
@@ -1966,7 +1966,7 @@ function slowlog_chart_measures(): array {
  * box's top value, since a rare true-max outlier can otherwise flatten the rest of the box.
  */
 function slowlog_get_stats_chart_object(string $chart_type, string $measure, array $scope_filter = array(), bool $hide_max = false): array {
-	$id = get_filter_request_var('logid');
+	$id = (int) get_filter_request_var('logid');
 
 	$description = db_fetch_cell_prepared('SELECT description
 		FROM plugin_slowlog
@@ -2049,7 +2049,7 @@ function slowlog_get_chart_object(string $chart_type, string $measure, array $sc
 		return array();
 	}
 
-	$id = get_filter_request_var('logid');
+	$id = (int) get_filter_request_var('logid');
 
 	$description = db_fetch_cell_prepared('SELECT description
 		FROM plugin_slowlog
